// src/app/shared/models/story.ts
export class Story {
    bts: {
        title: string;
        info: string;
    };
    description: {
        title: string;
        info: string;
    };
    fun: {
        title: string;
        info: string;
    };
    timeline: {
        endDate: any;
        startDate: any;
        startInfo: string;
        endInfo: string;
    };
}